+++ models/contacts_test.go
@@ -4,8 +4,8 @@ import (
 	"testing"
 
 	"github.com/nyaruka/gocommon/urns"
-	"github.com/nyaruka/goflow/flows"
-	"github.com/nyaruka/goflow/flows/engine"
+	"github.com/greatnonprofits-nfp/goflow/flows"
+	"github.com/greatnonprofits-nfp/goflow/flows/engine"
 	"github.com/nyaruka/mailroom/testsuite"
 	"github.com/shopspring/decimal"
 	"github.com/stretchr/testify/assert"